+++ src/orchard/pp.clj
@@ -123,7 +123,9 @@
   prefix."
   [coll]
   (if (record? coll)
-    [(str "#" (.getName (class coll)) "{") coll]
+    [(str "#" (if print/*short-record-names*
+                (.getSimpleName (class coll))
+                (.getName (class coll))) "{") coll]
     ;; If all keys in the map share a namespace and *print-
     ;; namespace-maps* is true, print the map using map namespace
     ;; syntax (e.g. #:a{:b 1} instead of {:a/b 1}). If the map is